refactor(interfaces): extract ingest and search DTOs to shared modules
The ingest and search pipeline DTOs are contracts shared by sources,
vector stores and bindings — they don't belong to any one layer. Move
them to neutral homes so every consumer depends on a common base:

- shared/ingest_types.py: IngestFile, IngestRequest, IngestContext
- shared/search_types.py: SearchContext, SearchParameters,
  SearchResult, SearchedDocument

Two files instead of one so the closures stay independent: ingest
shape changes don't drive search shape changes, and endpoint code
doesn't have to pull in ingest types just to do a query.

With the DTOs out of dataset_types/interfaces.py and
vector_stores/interfaces.py, the runtime cycle between those modules
disappears. The TYPE_CHECKING block and `from __future__ import
annotations` are no longer needed and have been removed; all
references are plain runtime imports now.
