+++ backend/syft_space/components/dataset_types/interfaces.py
@@ -1,12 +1,12 @@
-"""Dataset type interfaces and domain models.
+"""Dataset type interfaces.
 
 A ``BaseDatasetType`` is the binding of a ``BaseSource`` (data origin)
 and a ``BaseVectorStore`` (vector storage). Concrete bindings declare
 ``SOURCE_CLS`` and ``VECTOR_STORE_CLS`` as class attributes plus a
 ``split_config()`` classmethod that translates the flat user-facing
 configuration into the two per-axis configs; the default ``__init__``
-takes care of constructing each collaborator and exposing them as
-``self.source`` and ``self.vector_store``.
+constructs each collaborator and exposes them as ``self.source`` and
+``self.vector_store``.
 
 Lifecycle methods (``search``, ``healthcheck``, ``ingest``, ``delete``)
 delegate to the collaborators by default; bindings only override when
@@ -14,94 +14,20 @@ they need cross-axis policy (e.g. a source-defined allow-list applied
 at ingest time).
 """
 
-from __future__ import annotations
+from typing import Any, ClassVar
 
-from pathlib import Path
-from typing import TYPE_CHECKING, Any, ClassVar
-from uuid import UUID
-
-from pydantic import BaseModel, Field
-
-from syft_space.components.shared.domain_types import Context, HealthcheckResponse
-
-if TYPE_CHECKING:
-    from syft_space.components.sources.interfaces import BaseSource
-    from syft_space.components.vector_stores.interfaces import (
-        BaseVectorStore,
-        IngestableVectorStore,
-    )
-
-
-class SearchContext(Context):
-    """Context for search requests."""
-
-    dataset_id: UUID = Field(..., description="Unique identifier for the dataset")
-
-
-class IngestContext(Context):
-    """Context for ingestion requests."""
-
-    dataset_id: UUID = Field(..., description="Unique identifier for the dataset")
-
-
-class SearchParameters(BaseModel):
-    """Domain contract for search parameters."""
-
-    similarity_threshold: float = Field(
-        default=0.8, ge=0.0, le=1.0, description="Similarity threshold for matching"
-    )
-    limit: int = Field(
-        default=5, ge=1, description="Maximum number of results to return"
-    )
-    include_metadata: bool = Field(
-        default=True, description="Whether to include metadata in response"
-    )
-    extra_options: dict[str, Any] = Field(
-        default_factory=dict, description="Extra options for the search"
-    )
-
-
-class SearchedDocument(BaseModel):
-    """A single document from search results."""
-
-    document_id: str = Field(..., description="Unique identifier for the document")
-    content: str = Field(..., description="Content of the document")
-    metadata: dict[str, Any] = Field(
-        default_factory=dict, description="Document metadata"
-    )
-    similarity_score: float = Field(
-        ..., ge=0.0, le=1.0, description="Similarity score for the document"
-    )
-
-
-class SearchResult(BaseModel):
-    """Domain contract for search results."""
-
-    documents: list[SearchedDocument] = Field(
-        default_factory=list, description="List of searched documents"
-    )
-    metadata: dict[str, Any] = Field(
-        default_factory=dict, description="Additional search metadata"
-    )
-
-
-class IngestFile(BaseModel):
-    """Framework-agnostic file wrapper for ingestion."""
-
-    path: Path = Field(..., description="Local readable path")
-    filename: str = Field(..., description="Display filename")
-    file_size: int | None = Field(default=None, description="Size in bytes")
-    metadata: dict[str, Any] = Field(
-        default_factory=dict, description="Custom metadata"
-    )
-
-
-class IngestRequest(BaseModel):
-    """Domain contract for data ingestion."""
-
-    files: list[IngestFile] = Field(
-        default_factory=list, description="List of files to ingest"
-    )
+from syft_space.components.shared.domain_types import HealthcheckResponse
+from syft_space.components.shared.ingest_types import IngestContext, IngestRequest
+from syft_space.components.shared.search_types import (
+    SearchContext,
+    SearchParameters,
+    SearchResult,
+)
+from syft_space.components.sources.interfaces import BaseSource
+from syft_space.components.vector_stores.interfaces import (
+    BaseVectorStore,
+    IngestableVectorStore,
+)
 
 
 class BaseDatasetType:
